* 'Bug-fixes-1' of https://github.com/Eru111/tp: (22 commits)
  Update UG
  Fix multiple UG issues
  Fix for #201
  Fix UG bugs
  Fix formating
  Revert #223 to only fix grammar issues
  Revert "Fixes #232"
  Remove enhancement for #223
  Fix checkstyle issues
  Fixes #202
  Fixes #221
  Fixes #223
  Fixes #226
  Fixes #228
  Fixes #230
  Fixes #232
  Fixes #241
  Fixes #245 and #240
  Update AddCommandTest.java
  Fix `back` to show all persons
  ...

# Conflicts:
#	docs/UserGuide.md

+++ src/main/java/seedu/address/model/delivery/DeliveryList.java
@@ -162,7 +162,7 @@ public class DeliveryList {
     public void sortByCost() {
         internalList.sort((d1, d2) -> {
             if (isSameArchiveStatus(d1, d2)) {
-                return Float.compare(d1.getCost().asFloat(), d2.getCost().asFloat());
+                return Double.compare(d1.getCost().asDouble(), d2.getCost().asDouble());
             } else {
                 return d1.getArchive().value.compareTo(d2.getArchive().value);
             }
